+++ SystemC/C_Battery_model/source/Eq_8_12_RC_Components.h
@@ -1,1 +1,1 @@
-void RC_Components(int SOC, float *R_series, float *R_transient_S,float *C_transient_S, float *R_transient_L, float *C_transient_L);
+void RC_Components(float SOC, float *R_series, float *R_transient_S,float *C_transient_S, float *R_transient_L, float *C_transient_L);
